API-Fast-Foods-Inc/requirements.txt
astroid==1.6.5
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-JWT==0.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.2.0
pipenv==2018.7.1
pipfile==0.0.2
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.4
PyJWT==1.4.2
pylint==1.9.2
pytest==3.6.3
requests==2.19.1
six==1.11.0
toml==0.9.4
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
